libostree bundling fixes
Fix libostree bundling:
Ensure we do not Provide or Require libostree.
Move the shared library into flatpak-libs so flatpak always
depends on it.
Keep the shared library filename as libostree, but put it
under a private directory.  Renaming the file on disk did not
really do much since the dynamic linker and RPM work from the
soname.

Resolves: #1476905